backend/requirements.txt
torch
torchvision
timm
fastapi
uvicorn
mlflow
pillow
python-multipart
opencv-python-headless
